+++ CHANGELOG.md
@@ -3,6 +3,19 @@ All notable changes to this project will be documented in this file.
 
 The format is loosely based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [v0.3] - 2017-07-19
+### Added
+- Added QT-based GUI with controls for user input.
+- Added initial support for extended animation options. At the moment, it only allows you to set the threshold of lit LEDS in the Sparkle animation.
+- Added Colors::MixMode::NONE. This is currently used to disable an Overlay and/or flag it as uninitialized.
+- Added NEXT ColorAnimation as a shortcut for incrementing the currently active animation.
+
+### Removed
+- Removed GTK-based GUI
+
+### Known Issues
+- GUI does not support Overlays or multiple Sections.
+
 ## [v0.2] - 2017-05-01
 ### Added
 - Added Changelog file.